+++ CHANGELOG.md
@@ -1,9 +1,25 @@
+## 1.19.0
+### Features
+* add TypeScript definition file
+* update formatters API to have a way to return position after replace from function formatter
+* regex formatters and $.tracking_replace now accept function as replacement
+* update unix formatters to use new API so they work with command line
+* set exit to false if no login provided
+
+### Bugs
+* fix overtyping function [#409](https://github.com/jcubic/jquery.terminal/issues/409)
+* remove CR characters only for display
+* don't invoke onPosition change when calling position and don't change the value
+* fix clearing CR characters that was causing removal of empty lines [#411](https://github.com/jcubic/jquery.terminal/issues/411)
+
+
 ## 1.18.0
 ### Feature
 * looping regex formatters that replace until they don't match the regex
 * add tracking_replace to $.terminal namespace
 * $.terminal.syntax helper
 * new language for prism: "website" that handle html, javascript and css syntax
+
 ### Bugs
 * handle formatters that replace backspaces and characters before [#409](https://github.com/jcubic/jquery.terminal/issues/409)
 * fix broken < > & with cmd + prism [#410](https://github.com/jcubic/jquery.terminal/issues/410)